% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{add_knots.bmop}
\alias{add_knots.bmop}
\title{Add a new knot to a bmop object}
\usage{
add_knots.bmop(bmop, value, MARGIN = 1)
}
\arguments{
\item{bmop}{a bmop object}

\item{value}{the new knot to be added}

\item{MARGIN}{the dimension where the new knot has to be added}
}
\value{
a bmop object
}
\description{
Add a new knot to a bmop object
}

