% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{AIC.bmop}
\alias{AIC.bmop}
\title{Akaike Information Criteria for bmop}
\usage{
\method{AIC}{bmop}(object, data, corrected = F, ..., k = 2)
}
\arguments{
\item{object}{a bmop object}

\item{data}{data.frame, matrix, vector of observations}

\item{corrected}{logical}

\item{...}{see \code{\link[stats]{AIC}}}

\item{k}{penalization coefficients, see \code{\link[stats]{AIC}}}
}
\description{
Akaike Information Criteria for bmop
}

