% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-extra-functions.R
\name{as.bins}
\alias{as.bins}
\title{Create bins}
\usage{
as.bins(data, breaks = nclass.FD, ...)
}
\arguments{
\item{data}{data.frame or an object that will coerced to a data.frame.}

\item{breaks}{an integer, a vector of integer or a function, similar to
\code{hist}.}
}
\value{
a \code{bins} class object.
}
\description{
Create bins
}

