% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{as.function.bmop}
\alias{as.function.bmop}
\title{Convert an bmop object to a function}
\usage{
\method{as.function}{bmop}(x, MIN = 0, ...)
}
\arguments{
\item{x}{an bmop object}

\item{MIN}{non negative number, MIN value for evaluation of x}

\item{...}{compatibility with \code{\link{as.function}}}
}
\description{
Convert an bmop object to a function
}
\examples{
bmop<-bmop_fit(rnorm(100))
fun<-as.function(bmop)
fun(0)
fun(3)
plot(fun,from=-2,to=+2)
}

