% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{BIC.bmop}
\alias{BIC.bmop}
\title{Bayesian Information Criteria for bmop}
\usage{
\method{BIC}{bmop}(object, data, corrected = F, ...)
}
\arguments{
\item{object}{a bmop object}

\item{data}{data.frame, matrix, vector of observations}

\item{corrected}{logical}

\item{...}{see \code{\link[stats]{BIC}}}
}
\description{
Bayesian Information Criteria for bmop
}

