% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-estimation-functions.R
\name{bmop_fit.default}
\alias{bmop_fit.default}
\title{Estimation of bmop density or conditional density}
\usage{
\method{bmop_fit}{default}(data, density = TRUE, conditional = FALSE,
  Min = NULL, Max = NULL, ...)
}
\arguments{
\item{data}{data.frame, matrix, vector or object of class "histogram" or
"bins"}

\item{density}{logic, if \code{TRUE} a density is learned}

\item{conditional}{logic, if \code{TRUE} a conditional density is learned}

\item{Min}{vector of lower bounds}

\item{Max}{vector of upper bounds}

\item{...}{see \code{bmopPar}}
}
\value{
a bmop object
}
\description{
Estimation of bmop density or conditional density
}

