% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-estimation-functions.R
\name{bmop_fit.histogram}
\alias{bmop_fit.histogram}
\title{Estimation of bmop density or conditional density}
\usage{
\method{bmop_fit}{histogram}(data, conditional = F, Min = NULL,
  Max = NULL, bmop = NULL, ...)
}
\arguments{
\item{data}{\code{histogram} or \code{bins} object
the variables must be in the right order}
}
\description{
Estimation of bmop density or conditional density
}

