% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-estimation-functions.R
\name{bmop_fit}
\alias{bmop_fit}
\title{Estimation of bmop density or conditional density}
\usage{
bmop_fit(data, conditional = FALSE, Min = NULL, Max = NULL, bmop = NULL,
  ...)
}
\arguments{
\item{data}{data.frame, matrix, vector or object of class "histogram" or
"bins"}

\item{conditional}{logic, if \code{TRUE} a conditional density is learned}

\item{Min}{vector of lower bounds}

\item{Max}{vector of upper bounds}

\item{bmop}{a bmop object}

\item{...}{see \code{bmopPar}}
}
\value{
a bmop object, a density if \code{conditional=FALSE} or a
conditional density if \code{conditional=TRUE}. In the latter case the first
variable of the dataset is considered as the conditioned one and the rest of
the variables as the conditioning ones. If the dataset has only one variable
a normal density is generated discarding the value of \code{conditional}.
}
\description{
Estimation of bmop density or conditional density
}
\examples{
plot(bmop_fit(rnorm(100)))
plot(bmop_fit(hist(rnorm(100000))))
#############################
Data<-data.frame(rnorm(100),rexp(100))
bmop<-bmop_fit(Data)
plot(bmop)
#############################
X<-rnorm(100)
Y<-rnorm(100,mean=X)
Data<-data.frame(X,Y)
bmopPar(mle=TRUE)
bmopC<-bmop_fit(Data,conditional=TRUE)
bmopPar(mle=FALSE)
plot(bmopC)
}

