% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-estimation-functions.R
\name{bmop_fit.values}
\alias{bmop_fit.values}
\title{Estimation of bmop regression fucntions}
\usage{
\method{bmop_fit}{values}(data, Min = NULL, Max = NULL, bmop = NULL, ...)
}
\arguments{
\item{data}{\code{values} object
the variables must be in the right order}

\item{Min}{vector of lower bounds}

\item{Max}{vector of upper bounds}

\item{bmop}{a bmop object}

\item{...}{see \code{bmopPar}}
}
\value{
a bmop object
}
\description{
Estimation of bmop regression fucntions
}

