% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-estimation-functions.R
\name{bmopPar}
\alias{bmopPar}
\title{Set Rbmop parameters}
\usage{
bmopPar(...)
}
\arguments{
\item{...}{see details.}
}
\description{
Set appropriate global parameters to be used by \code{Rbmop}
functions . This is the appropriate way to change those parameters.
The use is similar to \code{par()} in package \code{graphics}.
}
\details{
This function set parameters used for estimation.The call \code{
          bmopPar()} just print the values of the paramters.


          \code{mle=FALSE}: logical. If use maximum-likelihood estimation of
          bmop coefficient, setting \code{mle=TRUE} just force
          \code{repMax=1}.

          \code{N=NA}: If present, the number of knots in every dimensions.
          Vector of positive integer, if needed values will be recycled.

          \code{order=3}: The order of the B-spline in every dimensions,
          vector of positive integer, if needed values will be recycled.

          \code{alpha=3}: The penalization
                        exponent to compute the number of knots. This is the
                        default method to compute the number of knots,
                        with the formula: \eqn{ floor(n^(1/alpha))}, where
                        \eqn{n} is the number of observations in the
                        dataset.
                        If \code{!is.na(N)} then the number of knots will be
                        set to \eqn{N^d} where \eqn{d} is the number of
                         dimensions
                        in the dataset (num. of variables).

         \code{knotsMethod="uniform"}:
          \code{"uniform"} or \code{"quantiles"} knots,
          how knots are computed by \code{\link{generate_knots}}.

          \code{k=2}: Coefficient of \code{AIC} (penalized likelihood),
          positive integer or \code{"BIC"} string. This is used by
          \code{\link{search_bmop}}.

           \code{toll=10^{-10}}: Tollerance for the increment of the likelihood in
                        the mle estimation of the coefficient.

          \code{repMax=100}: maximum number of iteration in the mle
                             estimation of the coefficient

         \code{MIN=10^{-10}}: This is not a learning parameter but instead
                              define the \code{MIN} parameter in the
                              evaluation of bmop object. Observe that some
                              functions like \code{\link{logLik}} or
                              \code{plot},
                              set this parameter independently.

         \code{autoReduce=200}: This value set the maximum dimension of an
                                accepted dataset as raw-data, for larger
                                dataset, functions \code{\link{bmop_fit}}
                                 and
                                \code{\link{search_bmop}}
                                 will be applied over the
                                reduced bins (histogram). Setting it to
                                \code{Inf} disable this features.
}

