% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3-manipulation-functions.R
\name{clean.bmop}
\alias{clean.bmop}
\title{Clean a bmop object}
\usage{
clean.bmop(object)
}
\arguments{
\item{object}{a bmop object}
}
\value{
a bmop object
}
\description{
remove extra informations from bmop object
}

