% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4-comparison-functions.R
\name{compare.bmop}
\alias{compare.bmop}
\title{Compare bmop with true density}
\usage{
compare.bmop(object, dtrue, measure = "MSE", method = "grid",
  densit = dtrue, N = 100, ...)
}
\arguments{
\item{object}{bmop object}

\item{dtrue}{function}

\item{measure}{string, \code{"MSE"}, \code{"MAE"}, \code{"MAX"}}

\item{method}{string, \code{"grid"} or \code{"montecarlo"}}

\item{densit}{function or string \code{"uniform"}}

\item{N}{positive integer, number of points of comparison}

\item{...}{optional arguments to be passed to \code{dtrue}}
}
\description{
Compare bmop with true density
}
\examples{
data<-rnorm(200)
bmopE<-estimate_bmop(data)
bmopS<-search_bmop(data)
compare.bmop(bmopE,dtrue=dnorm)
compare.bmop(bmopS,dtrue=dnorm,method="montecarlo")
}

