% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4-comparison-functions.R
\name{comparison_plot}
\alias{comparison_plot}
\title{Plot several bmops and true density}
\usage{
comparison_plot(bmop.list, dtrue = NULL, colors = NULL, lwd = 3,
  type = "l", type.true = "l", col.true = "red",
  names.bmop = 1:length(bmop.list), legend.display = T,
  legend.pos = "topleft", file = NULL, ylim = c(0, 1), ...)
}
\arguments{
\item{bmop.list}{list of bmop objects (1 dimensional)}

\item{dtrue}{function, true density function}

\item{colors}{NULL or vector of colors}

\item{lwd}{graphical parameter see \code{\link{par}}}

\item{type}{graphical parameter see \code{\link{par}}}

\item{type.true}{as \code{type}, but just for true density plot}

\item{col.true}{as \code{col} graphical parameter but for dtrue density plot}

\item{names.bmop}{vector for the names showed in legend}

\item{legend.display}{logical, to show or not the legend}

\item{legend.pos}{string for the position of the legend
\code{"top"}, \code{"left"},
\code{"topright"}, etc.}

\item{...}{more parameters to be passed to \code{dtrue}}
}
\value{
invisible()
}
\description{
Plot several bmops and true density
}
\examples{
data<-rnorm(200)
bmopE<-estimate_bmop(data)
bmopS<-search_bmop(data)
comparison_plot(list(bmopE,bmopS),true=dnorm,
                 names.bmop=c("Fast","AIC-search"))
}

