% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-estimation-functions.R
\name{cond_estimate_bmop}
\alias{cond_estimate_bmop}
\title{Fast Estimation of bmop conditional density}
\usage{
cond_estimate_bmop(data, bmop = NULL, Min = NULL, Max = NULL, ...)
}
\arguments{
\item{data}{data.frame, matrix or vector,
the variables must be in the right order (the columns of data)}

\item{Min}{real or vector}

\item{Max}{real or vector}

\item{...}{see \code{bmopPar}}
}
\value{
A bmop object, the aproximations of \eqn{f}.
}
\description{
Aproximation of a conditional density \eqn{f(x| x_1,\ldots,x_n)}
}
\examples{
require("MASS")
X<-rnorm(100)
Y<-rnorm(100,mean=X)
data<-data.frame(Y,X)
condbmop<-estimate_bmop(data=data)
plot(condbmop,N=20,persp=TRUE)
}

