% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-estimation-functions.R
\name{cond_mle_fit_bmop}
\alias{cond_mle_fit_bmop}
\title{Maximum conditional-likelihood estimator for the coefficient of B-spline}
\usage{
cond_mle_fit_bmop(data, toll = 10^(-1), repmax = 50, fresh.start = TRUE,
  bmop = NULL, Min = NULL, Max = NULL, return.bmop = TRUE, ...)
}
\arguments{
\item{data}{data.frame, matrix or vector first column is the
x variable and the following columns are for
conditioning variable \eqn{x_1,...,x_n}}

\item{toll}{numeric tollerance parameter for the loglik increase}

\item{repmax}{positive integer maximum number of repetitions}

\item{fresh.start}{logical if TRUE the cotrol points of mop are set to uniform}

\item{bmop}{bmop object}

\item{Min}{real or vector}

\item{Max}{real or vector}

\item{...}{see \code{bmopPar}}
}
\value{
if return.mop=TRUE a bmop object, the aproximations of \eqn{f},
else a data.frame of control points.
}
\description{
Aproximation of a conditional density \eqn{f(x| x_1,\ldots,x_n)} by maximum
cond-likelihood, the type of mop, knots and order is fixed by the user
}

