% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-estimation-functions.R
\name{cond_search_bmop}
\alias{cond_search_bmop}
\title{Greedy penalized conditional log-likelihood search}
\usage{
cond_search_bmop(data, k = Rbmop::bmopPar()$k, corrected = FALSE,
  knotsMethod = Rbmop::bmopPar()$knotsMethod, ...)
}
\arguments{
\item{data}{data.frame, matrix or vector, the variables
must be in the right order (the columns of data)}

\item{k}{positive number or \code{"BIC"}}

\item{corrected}{logical}
}
\value{
A bmop object, the aproximations of \eqn{f}.
}
\description{
Aproximation of a conditional density density \eqn{f(x|x_1,\ldots,x_n)}
}
\examples{
X<-rnorm(100)
Y<-rnorm(100,mean=X)
data<-data.frame(Y,X)
condbmop<-cond_search_bmop(data=data,k="BIC")
plot(condbmop,N=20,persp=TRUE)
}

