% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{deboor}
\alias{deboor}
\title{DeBoor algorithm
as in
\url{http://cse.taylor.edu/~btoll/s99/424/res/ucdavis/CAGDNotes/
Deboor-Cox-Calculation/Deboor-Cox-Calculation.html}
knots
sequence must be of type \eqn{u_1=u_2=...=u_(k)<u_r<...<u_s<u_(k+1)=..=u_n}
 see \code{\link{generate_knots}}}
\usage{
deboor(t, k, knots, ctr, MIN = 10^(-10))
}
\arguments{
\item{t}{numeric}

\item{k}{positive integer}

\item{knots}{vector}

\item{ctr}{vector}

\item{MIN}{numeric}
}
\value{
numeric value of the computed bspline at point t
}
\description{
DeBoor algorithm
as in
\url{http://cse.taylor.edu/~btoll/s99/424/res/ucdavis/CAGDNotes/
Deboor-Cox-Calculation/Deboor-Cox-Calculation.html}
knots
sequence must be of type \eqn{u_1=u_2=...=u_(k)<u_r<...<u_s<u_(k+1)=..=u_n}
 see \code{\link{generate_knots}}
}

