% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-estimation-functions.R
\name{define_bmop}
\alias{define_bmop}
\title{define bmop, internal function}
\usage{
define_bmop(bmop = NULL, data = NULL, Max = NULL, Min = NULL,
  N = get(".bmopPars", envir = .bmopenv)$N, order = get(".bmopPars", envir =
  .bmopenv)$order, alpha = get(".bmopPars", envir = .bmopenv)$alpha,
  method = get(".bmopPars", envir = .bmopenv)$knotsMethod)
}
\description{
define bmop, internal function
}

