% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{delta}
\alias{delta}
\title{return an array with the B-spline basis computed over the points in \code{x}
CHANGE ITTTTTT
Internal function}
\usage{
delta(bmop, x, ...)
}
\arguments{
\item{bmop}{bmop object}

\item{x}{numeric}

\item{...}{oprional parameters}
}
\value{
array with the values of the bspline basis corresponding to bmop
}
\description{
return an array with the B-spline basis computed over the points in \code{x}
CHANGE ITTTTTT
Internal function
}

