% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{dim.bmop}
\alias{dim.bmop}
\title{Dimension of bmop}
\usage{
\method{dim}{bmop}(x)
}
\arguments{
\item{x}{a bmop object}
}
\value{
the number of free parameters of \code{x}
}
\description{
Dimension of bmop
}

