% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4-comparison-functions.R
\name{envelope_plot}
\alias{envelope_plot}
\title{Envelop of estimated bmop}
\usage{
envelope_plot(n = 100, N = 50, rtrue = rnorm, fun = bmop_fit,
  dtrue = dnorm, lwd = 3, type = "l", col.true = "red", ...)
}
\arguments{
\item{n}{number of bmop densities to learn from different datasets}

\item{N}{number of observations in every dataset}

\item{rtrue}{function to generate samples, see \code{\link{rnorm}}}

\item{fun}{a learning function of bmop as \code{\link{bmop_fit}}}

\item{dtrue}{true density function, see \code{\link{dnorm}}}

\item{lwd}{graphical par}

\item{type}{graphical par}

\item{col.true}{the color for the true density}

\item{...}{additional parameters to be passed to \code{rtrue} and
\code{dtrue}}
}
\value{
\code{invisible() }
}
\description{
This function plots various bmop estimations, from differents datasets.
}
\examples{
envelope_plot(n=50,N=50,rtrue=rexp,dtrue=dexp)
}

