% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-estimation-functions.R
\name{estimate_bmop}
\alias{estimate_bmop}
\title{Fast Estimation of bmop density}
\usage{
estimate_bmop(data, bmop = NULL, conditional = F, Min = NULL,
  Max = NULL, ..., D = NULL)
}
\arguments{
\item{data}{data.frame, matrix or vector,
the variables must be in the right order (the columns of data)}

\item{Min}{real or vector}

\item{Max}{real or vector}

\item{...}{see \code{bmopPar}}

\item{D}{internal parameter}
}
\value{
A bmop object, the aproximations of \eqn{f}.
}
\description{
Aproximation of a density \eqn{f(x_1,\ldots,x_n)}
}
\examples{
require("MASS")
data<-mvrnorm(n=50,mu=c(0,0),Sigma=diag(x=c(1,1),nrow = 2,ncol = 2))
bmop<-estimate_bmop(data=data)
plot(bmop,N=20,persp=TRUE)

data<-rnorm(n=1000)
bmop<-estimate_bmop(data=data)
plot(bmop)
summary(bmop)
compare.bmop(object=bmop,dtrue=rnorm,N=1000)
}

