% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/5-example-bmop.R
\name{Examples_bmop}
\alias{Examples_bmop}
\alias{ex_bmop_gaussian2Mixture}
\alias{ex_bmop_gaussian3Mixture}
\alias{ex_bmop_gaussianBetaGamma}
\title{Examples of bmop density estimations}
\usage{
ex_bmop_gaussian2Mixture(N = 1000, m1 = -3, m2 = 0, lambda = 0.5)

ex_bmop_gaussian3Mixture(N = 1000, m1 = -3, m2 = 0, m3 = +3,
  lambda = c(1, 1, 1))

ex_bmop_gaussianBetaGamma(N = 1000, m1 = -3, m2 = 2, m3 = 3,
  lambda = c(1, 1, 1))
}
\arguments{
\item{N}{positive integer, the number of observations}

\item{m1,m2,m3}{location parameters}

\item{lambda}{mixing coefficient, vector or double}
}
\value{
All the functions return an invisible list contating the
generated dataset, the estimated bmops and the true density function.
}
\description{
Various examples to show the capabilities of bmop estimations.
}
\details{
This functions generate datasets of \code{N} observations.
The function \code{bmop_fit} is then used to estimate the density.
A comparative plot is then returned.
 \describe{
 \item{\code{ex_bmop_gaussian2Mixture}}{\code{lambda}-mixture
 of two Gaussian densities with unitary variance
 and means \code{m1} and \code{m3}.}
 \item{\code{ex_bmop_gaussian3Mixture}}{\code{lambda}-mixture
 of three Gaussian densities with unitary variance
 and means \code{m1}, \code{m2} and \code{m3}.}
 \item{\code{ex_bmop_gaussianBetaGamma}}{\code{lambda}-mixture
 of a Gaussian density with unitary variance
 and mean \code{m1}, a Beta density with \code{shape1=2 shape2=5 ncp=m2}  and
 a Gamma density with \code{shape=9 scale=m3/9}.}
 }
}

