% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-extra-functions.R
\name{fix_data}
\alias{fix_data}
\title{Fix Data}
\usage{
fix_data(data)
}
\arguments{
\item{data,}{vector, array, data.frame, matrix}
}
\value{
if \code{is.null(dim(data))} then the function set
\code{dim(data)<-c(length(data),1)} and return \code{data}.
}
\description{
Fix Data
}

