% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{generate_knots}
\alias{generate_knots}
\title{Generate the sequence of knots for bmop objects}
\usage{
generate_knots(data = NULL, N = 5, method = "uniform", Min = NULL,
  Max = NULL)
}
\arguments{
\item{data}{data.frame matrix vector}

\item{N}{positive integer, the number of knots}

\item{method}{string "uniform" or "quantiles"}

\item{Min}{numeric}

\item{Max}{numeric}
}
\description{
Generate the sequence of knots for bmop objects
}
\examples{
data<-rnorm(200)
bmopE<-bmop_fit(data)
bmopS<-search_bmop(data)
plot(bmopE)
points(bmopS,type="l",col="red")
}

