% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{integration_constants}
\alias{integration_constants}
\title{Integral of the B-spline base}
\usage{
integration_constants(bmop, append.extra = F)
}
\arguments{
\item{bmop}{bmop object}

\item{append.extra}{logical, if TRUE then the original bmop is returned,
with the integration constants in \code{bmop$extra$C}}
}
\value{
array of the values of the integrals of the bspline base
}
\description{
Compute the integrals of the bspline basis over all the domain.
}
\seealso{
integrate
}

