% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{is.bmop}
\alias{is.bmop}
\title{Check if an object's class is bmop}
\usage{
is.bmop(object)
}
\arguments{
\item{object}{an R object}
}
\value{
logical, TRUE if "bmop" is present among class(object),
 FALSE otherwise
}
\description{
Check if an object's class is bmop
}
\details{
Just a call of \code{inherits(object,"bmop")}
}
\examples{
bmop<-bmop_fit(rexp(100))
is.bmop(bmop)
}

