% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4-comparison-functions.R
\name{KL.bmop}
\alias{KL.bmop}
\title{Kullback-Leibler (KL) divergence between bmop and true density}
\usage{
KL.bmop(object, dtrue, ...)
}
\arguments{
\item{object}{bmop object}

\item{dtrue}{function}

\item{...}{optional arguments to be passed to \code{dtrue}}
}
\value{
Numeric, the KL divergence.
}
\description{
Kullback-Leibler (KL) divergence between bmop and true density
}
\examples{
data<-rnorm(200)
bmop1<-bmop_fit(data)
bmopPar(mle=TRUE)
bmop2<-bmop_fit(data)
KL.bmop(bmop1,dtrue=dnorm)
KL.bmop(bmop2,dtrue=dnorm)
}

