% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-extra-functions.R
\name{locate}
\alias{locate}
\title{Position of an element into a sorted vector}
\usage{
locate(a, v)
}
\arguments{
\item{a}{value}

\item{v}{vector in ascending order (or it will be ordered)}
}
\value{
positive integer, the position i of the first element
of \eqn{v} (\eqn{v_i}) such that \eqn{v_i<=a<v_{i+1}}
}
\description{
Position of an element into a sorted vector
}

