% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{logLik.bmop}
\alias{logLik.bmop}
\title{Log-Likelihood of bmop object}
\usage{
\method{logLik}{bmop}(object, data, ...)
}
\arguments{
\item{object}{a bmop object}

\item{data}{matrix, data.frame or vector of observation}

\item{...}{some methods require additional arguments,
see \code{\link[stats]{AIC}}}
}
\description{
Log-Likelihood of bmop object
}
\seealso{
AIC.bmop
}

