% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{lower.bmop}
\alias{lower.bmop}
\title{Lower Limit of bmop}
\usage{
lower.bmop(object)
}
\arguments{
\item{object}{a bmop object}
}
\value{
numeric value
}
\description{
Lower Limit of bmop
}

