% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3-manipulation-functions.R
\name{marginalize.bmop}
\alias{marginalize.bmop}
\title{Marginalize a bmop}
\usage{
marginalize.bmop(object, MARGIN = 1, ...)
}
\arguments{
\item{object}{a bmop object}

\item{MARGIN}{positive integer or vector of positive integer,
the dimensions that has to be marginalized}
}
\value{
a bmop object over a space of dimension \code{length(MARGIN)}
the result of integrating over the \code{-MARGIN}
}
\description{
marginalize a bmop object, that is integrate a bmop multivariate density
over some of its dimension.
}
\examples{
data<-data.frame(rnorm(100),rnorm(100))
bmop2d<-estimate_bmop(data)
bmop1d<-estimate_bmop(data[,1])
bmop1dmargin<-marginalize.bmop(bmop2d,MARGIN=1)
comparison_plot(list(bmop1d,bmop1dmargin),true=dnorm,
names.bmop=c("direct est.","marginalized"))
}

