% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-estimation-functions.R
\name{mle_fit_bmop}
\alias{mle_fit_bmop}
\title{Maximum likelihood estimator for the coefficient of B-spline}
\usage{
mle_fit_bmop(data, toll = 10^(-2), repmax = 50, fresh.start = TRUE,
  return.bmop = TRUE, bmop = NULL, Min = NULL, Max = NULL, ...)
}
\arguments{
\item{data}{data.frame, matrix or vector,
the variable must be in the right order (the columns of data)}

\item{toll}{numeric tollerance parameter for the loglik increase}

\item{repmax}{positive integer maximum number of repetitions}

\item{fresh.start}{logical if \code{TRUE} the cotrol points of mop are set
to uniform}

\item{return.bmop}{logical, if \code{TRUE} the function return a bmop object}

\item{bmop}{bmop object (optional)}

\item{...}{see \code{bmopPar}}
}
\value{
if \code{return.bmop=TRUE} a bmop object,
the aproximations of \eqn{f}, else a data.frame of control points.
}
\description{
Aproximation of a density \eqn{f(x_1,\ldots,x_n)} by maximum likelihood,
the type of mop, knots and order is fixed by the user
}
\examples{
require("MASS")
data<-mvrnorm(n=50,mu=c(0,0),Sigma=diag(x=c(1,1),nrow = 2,ncol = 2))
bmop<-new_bmop(knots=generate_knots(data),order=c(3,3))
bmop<-mle_fit_bmop(data=data,bmop=bmop,return.bmop=TRUE)
plot(bmop,N=20,persp=TRUE)
}

