% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-extra-functions.R
\name{multiplicity}
\alias{multiplicity}
\title{The numbers of occurency of an elemnent into a vector}
\usage{
multiplicity(x, v)
}
\arguments{
\item{x}{value}

\item{v}{vector}
}
\value{
numeric value giving the number of times an element appears into
the vector
}
\description{
The numbers of occurency of an elemnent into a vector
}

