% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3-manipulation-functions.R
\name{normalize.bmop}
\alias{normalize.bmop}
\title{Normalize a bmop}
\usage{
normalize.bmop(object, ...)
}
\arguments{
\item{object}{a bmop object}
}
\value{
a bmop object proportional to object,
but such that integrates to one.
}
\description{
Normalize a bmop object, makes it integrate to one.
}

