% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{points.bmop}
\alias{points.bmop}
\title{plot points from bmop}
\usage{
\method{points}{bmop}(x, N = 100, ...)
}
\arguments{
\item{x}{bmop object}

\item{N}{number of points to plot}

\item{...}{graphical parameters see \code{\link{par}}}
}
\description{
plot points from bmop
}
\examples{
data<-rnorm(200)
bmopE<-estimate_bmop(data)
bmopS<-search_bmop(data)
plot(bmopE)
points(bmopS,type="l",col="red")
}
\seealso{
plot.bmop
}

