% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{print.bmop}
\alias{print.bmop}
\title{Print bmop objects}
\usage{
\method{print}{bmop}(x, ...)
}
\arguments{
\item{x}{bmop object}

\item{...}{see \code{\link{print}}}
}
\description{
Print bmop objects
}

