% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3-manipulation-functions.R
\name{put_evidence.bmop}
\alias{put_evidence.bmop}
\title{Put evidence on a conditional bmop}
\usage{
put_evidence.bmop(object, evidence, evd.pos = NULL, MIN = 0,
  normalize = FALSE, ...)
}
\arguments{
\item{object}{a bmop object}
}
\value{
a bmop object, the result of imposing some evidence
}
\description{
Normalize a bmop object, makes it integrate to one.
}

