% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-extra-functions.R
\name{sampler_MH}
\alias{sampler_MH}
\title{Metropolis-Hasting Sampler}
\usage{
sampler_MH(N, d, densit, h = 3, M = 1000, xstart = NULL, max = +Inf,
  min = -Inf)
}
\arguments{
\item{N}{positive integer, the number of observations}

\item{d}{integer >1, the dimension of the random vector}

\item{densit}{function, the (multi valued) density function}

\item{h}{positive integer, the jumping width between recorded observations}

\item{M}{positive integer, the heating phase length}

\item{xstart}{d-dimensional vector (or NULL) the starting point of the
Markov chain}

\item{max}{numeric, truncation parameter}

\item{min}{numeric, truncation parameter}
}
\value{
matrix of observations
}
\description{
Metropolis-Hasting Sampler
}
\examples{
sample1<-sampler_MH(100,1,dnorm)
hist(sample1)
plot(bmop_fit(sample1))
sample2<-sampler_MH(100,1,dnorm,max=0.5,min=-1)
hist(sample2)
}

