% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4-comparison-functions.R
\name{squareError.bmop}
\alias{squareError.bmop}
\title{Square Error between bmop and true density}
\usage{
squareError.bmop(object, dtrue, lowerLimit = NULL, upperLimit = NULL, ...)
}
\arguments{
\item{object}{bmop object}

\item{dtrue}{function}

\item{...}{optional arguments to be passed to \code{dtrue}}
}
\value{
Numeric, the square error between bmop denisty and true density
}
\description{
Square Error between bmop and true density
}
\examples{
data<-rnorm(200)
bmop1<-bmop_fit(data)
bmopPar(mle=TRUE)
bmop2<-bmop_fit(data)
squareError.bmop(bmop1,dtrue=dnorm)
}

