% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{upper.bmop}
\alias{upper.bmop}
\title{Upper Limit of bmop}
\usage{
upper.bmop(object)
}
\arguments{
\item{object}{a bmop object}
}
\value{
numeric value
}
\description{
Upper Limit of bmop
}

