% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0-extra-functions.R
\name{as.bins}
\alias{as.bins}
\title{Bins grouping}
\usage{
as.bins(data, breaks = nclass.FD, ...)
}
\arguments{
\item{data}{a dataset}

\item{breaks}{function or positive integer}

\item{...}{additional parameters}
}
\value{
an object \code{bins}, the data are grouped into bins uniformly
}
\description{
This function extend the histogram class for multi-dimensional datasets
}

