% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2-estimation-functions.R
\name{bmop_fit.data.frame}
\alias{bmop_fit.data.frame}
\title{Estimation of bmop density or conditional density}
\usage{
\method{bmop_fit}{data.frame}(data, density = TRUE, conditional = F,
  Min = NULL, Max = NULL, bmop = NULL, ...)
}
\arguments{
\item{data}{data.frame, matrix or vector
the variables must be in the right order (the columns of data)}

\item{density}{logic, if \code{TRUE} a density is learned}

\item{conditional}{logic, if \code{TRUE} a conditional density is learned}

\item{Min}{vector of lower bounds}

\item{Max}{vector of upper bounds}

\item{bmop}{a bmop object}

\item{...}{see \code{bmopPar}}
}
\value{
a bmop object
}
\description{
Estimation of bmop density or conditional density
}

