% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4-comparison-functions.R
\name{compare.bmop}
\alias{compare.bmop}
\title{Compare bmop with true density}
\usage{
compare.bmop(object, dtrue, measure = "MSE", ...)
}
\arguments{
\item{object}{bmop object}

\item{dtrue}{function}

\item{measure}{string, \code{"MSE"}, \code{"MAE"}, \code{"MAX"}}

\item{...}{optional arguments to be passed to \code{dtrue}}
}
\description{
Compare bmop with true density
}
\examples{
data<-rnorm(200)
bmop1<-bmop_fit(data)
bmopPar(mle=TRUE)
bmop2<-bmop_fit(data)
compare.bmop(bmop1,dtrue=dnorm)
compare.bmop(bmop2,dtrue=dnorm,method="montecarlo")
}

