% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/4-comparison-functions.R
\name{envelope_plot}
\alias{envelope_plot}
\title{Envelop of estimated bmop}
\usage{
envelope_plot(n = 100, N = 50, rtrue = rnorm, fun = bmop_fit,
  dtrue = dnorm, lwd = 3, type = "l", col.true = "red", ...)
}
\value{
invisible()
}
\description{
This function plots various bmop estimations, from differents datasets.
}
\examples{
envelope_plot(n=50,N=50,rtrue=rexp,dtrue=dexp)
}

