% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{evaluate.bmop}
\alias{evaluate.bmop}
\title{Evaluation of a bmop object}
\usage{
evaluate.bmop(x, object, MIN = bmopPar()$MIN)
}
\arguments{
\item{x}{numeric, vector, matrix}

\item{object}{bmop object}

\item{MIN}{numeric}
}
\value{
Numeric value (values) of the computed bspline at point
 (points) \code{x},
        deBoor algorithm is used.
}
\description{
Evaluation of a bmop object
}
\examples{
bmop<-bmop_fit(rnorm(100))
evaluate.bmop(0,bmop)
evaluate.bmop(c(-1,0,+1),bmop)
}

