% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{integrate.bmop}
\alias{integrate.bmop}
\title{Integrate bmop object over the support}
\usage{
integrate.bmop(object)
}
\arguments{
\item{object}{bmop object}
}
\description{
Integrate bmop object over the support
}
\examples{
data<-rnorm(200)
bmop<-estimate_bmop(data)
integrate.bmop(bmop)
}

