% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{mean.bmop}
\alias{mean.bmop}
\title{Mean value for a bmop density}
\usage{
\method{mean}{bmop}(x, ...)
}
\arguments{
\item{x}{a bmop object}

\item{...}{for compatibility with \code{\link{mean}}}
}
\value{
numeric value, the mean of the bmop density
}
\description{
Not very suited for high dimension
}
\examples{
bmop<-bmop_fit(rnorm(100))
mean(bmop)
}

