% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{new_bmop}
\alias{new_bmop}
\title{New bmop object}
\usage{
new_bmop(knots, order, ctrpoints = NULL, simple = TRUE, nk = FALSE)
}
\arguments{
\item{knots}{list of numeric vector}

\item{order}{vector}

\item{ctrpoints}{array}

\item{simple}{logical}

\item{nk}{logical}
}
\value{
An object of class bmop, a list with elements \code{order}, \code{knots}, \code{ctrpoints}
}
\description{
Constructor of bmop object.
}

