% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{plot.bmop}
\alias{plot.bmop}
\title{Plot of bmop object}
\usage{
\method{plot}{bmop}(x, N = 1000, type = "l", contour = TRUE,
  persp = FALSE, file = NULL, MIN = 0, ...)
}
\arguments{
\item{x}{bmop object}

\item{N}{positive integer}

\item{type}{graphical parameter, see \code{\link{par}}}

\item{contour}{graphical parameter, see \code{\link{par}}}

\item{persp}{logical}

\item{file}{optional file name, where to save a pdf copy of the plot}

\item{MIN}{parameter of \code{\link{evaluate.bmop}}}

\item{...}{arguments to be passed to methods, such as graphical parameters
(as in plot.default)}
}
\value{
plot of 1-d or 2-d (contour and perspective) bmop, an emty plot and
a warning message are returned if the bmop object has more than 2 dimension
}
\description{
Plot of bmop object
}
\seealso{
\link{points.bmop}
}

