% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1-base-functions.R
\name{points.bmop}
\alias{points.bmop}
\title{Plot points from bmop}
\usage{
\method{points}{bmop}(x, N = 100, ...)
}
\arguments{
\item{x}{bmop object}

\item{N}{number of points to plot}

\item{...}{graphical parameters see \code{\link{par}}}
}
\description{
Plot points from bmop
}
\details{
As \code{\link{points}}, this functions provide a way to add the
plot of a bmop to an existing plot.
}
\examples{
data<-rnorm(200)
bmopE<-bmop_fit(data)
bmopS<-search_bmop(data)
plot(bmopE)
points(bmopS,type="l",col="red")
}
\seealso{
\code{\link{plot.bmop}}.
}

