% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/3-manipulation-functions.R
\name{put_evidence.bmop}
\alias{put_evidence.bmop}
\title{Put evidence on a conditional bmop}
\usage{
put_evidence.bmop(object, evidence, evd.pos = NULL, MIN = 0,
  normalize = FALSE, ...)
}
\arguments{
\item{object}{a bmop object}

\item{evidence}{the value of evidence}

\item{evd.pos}{the position of evidence}

\item{MIN}{the MIN value as in \code{bmopPar}}

\item{normalize}{logical, if \code{TRUE} the final bmop object will
be normalized (usually it is not needed since this function is applied to
conditional densities)}
}
\value{
a bmop object, the result of imposing some evidence
}
\description{
Normalize a bmop object, makes it integrate to one.
}

